app/requirements.txt
flask>=2.0
psycopg2-binary>=2.9